feat(openssl-vuln-poc): implement morello PoCs
Run the OpenSSL Proof-of-concepts remotely via SSH.

In order to do this, we can just run the OpenSSL commands
through `ssh`. We can't call the OpenSSL commands directly,
since we need a shell to capture the SIGABRT or SIGPROT signals
that the vulnerable OpenSSL processes throw. The shell will
print a nice error message for us, then convert the signal to
an exit code (`134` for SIGABRT and `162` for `SIGPROT`).

Currently, I've set up the test cases to work with a local Docker
image, but you should be able to modify the username/port/IP to point
to a real Morello box if you want.