fix(openssl-vuln-poc): always return `stdin` prop
Currently, the OpenSSL runTest() commands returns a `client.stdin`
and a `server.stdin` property if the server or the client command
fails.

(side-note, `stdin` is a misnomer, since the property actually contains
 the command args used to run the commands. The real stdin is not
 returned)

However, since the app uses the `server.stdin` property, this causes
issues if the property doesn't exist.

This PR changes the code to always return the `client.stdin` and
`server.stdin` properties. I've also added tests for them, and updated
the mocks to have these properties.