fix(server): run server setup() once only
Run the MorelloOpenSSLTestCase#setup() only once to avoid
race-conditions, and to improve performance.

We automatically run it when the server starts up.

While the setup() code is still running, and API requests to the server
return a HTTP 503 Service Unavailable error, with a `Retry-After`
header telling the users to try again in a few seconds.

The error message also contains a very very rough ETA on how long the
Morello OpenSSL Test Cases will take to setup.