fix(app): prevent unnecessary testcase SWR queries
The way we're using SWR is not correct for the Morello test cases.
Currently, the app is constantly make Morello requests.

Clicking a button normally results in an API request being made for
the old button (which is unused), then the new config.

We should be making a query to server if-and-only-if a button
is pressed.

This goes against what SWR is meant for, but we can fake it by:
  - useSWRImmutable, so that SWR doesn't automatically rerender when
    - The page is refocussed
    - The page is resized
    - The page goes offline/online
  - By calling `mutate()` with the `revalidate: false` option.