feat(app): combine API server and app server
Combine the API server and the app server so that they both run in a
single express instance on a single port.

Benefits:
  - Simpler code
  - Single port to open
  - Single process to watch
  - Better logging
  - Better HTTP handling

The main reason to make this change is that now the server's HTTP error
responses are correctly being sent to the app, so that they can be
handled.